Hide release indicators when there's no data (e.g. for kube)
